Add filtering methods to Mesh and PointCloud
FilterMesh and FilterCloud can be used to select a sub-set of faces or
vertices that we want to keep. The motivation behind these is to enable
tools like cropping or outlier removal directly on the draco
datastructure.

FilterCloud is very similar to ApplyPointIdDeduplication and I was
originally considering sharing the implementation here. From an outside
POV, they have slightly different purposes though:
- FilterCloud selects which points to keep, discarding the rest
- ApplyPointIdDeduplication discards duplicated points, assuming all old
points have a (deduplicated) target new point